backend/requirements.txt
Flask==3.1.3
Flask-Session==0.8.0
flask-cors==6.0.2
oracledb==1.4.2
google-generativeai==0.3.2
pandas
Werkzeug==3.1.8
requests==2.31.0
python-dotenv
PyMySQL==1.1.0
psycopg2-binary==2.9.11
gunicorn